etape_06_langchain/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
